+++ R/model_performance.bayesian.R
@@ -38,6 +38,15 @@
 #'   )
 #'   model_performance(model)
 #' }
+#'
+#' if (require("BayesFactor")) {
+#'   model <- generalTestBF(carb ~ am + mpg, mtcars)
+#'
+#'   model_performance(model)
+#'   model_performance(model[3])
+#'
+#'   model_performance(model, average = TRUE)
+#' }
 #' @seealso \link{r2_bayes}
 #' @references Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, The American Statistician, 1-6.
 #'
@@ -117,3 +126,93 @@ model_performance.stanreg <- function(model, metrics = "all", verbose = TRUE, ..
 
 #' @export
 model_performance.brmsfit <- model_performance.stanreg
+
+
+#' @export
+#' @inheritParams r2_bayes
+#' @rdname model_performance.stanreg
+model_performance.BFBayesFactor <- function(model, metrics = "all", verbose = TRUE,
+                                            average = FALSE, prior_odds = NULL, ...) {
+  if (all(metrics == "all")) {
+    metrics <- c("R2", "SIGMA")
+  }
+
+  mi <- insight::model_info(model)
+
+  out <- list()
+  attri <- list()
+
+  if ("R2" %in% c(metrics)) {
+    r2 <- r2_bayes(model, average = average, prior_odds = prior_odds)
+    attri$r2_bayes <- attributes(r2) # save attributes
+
+    # Format to df then to list
+    r2_df <- as.data.frame(t(as.numeric(r2)))
+    names(r2_df) <- gsub("_Bayes", "", names(r2), fixed = TRUE)
+    out <- append(out, as.list(r2_df))
+  }
+
+
+  if ("SIGMA" %in% toupper(metrics)) {
+    sig <- suppressMessages(.get_sigma_bfbayesfactor(model, average = average, prior_odds = prior_odds))
+    out$Sigma <- bayestestR::point_estimate(sig, "median")[[1]]
+  }
+
+
+  out <- as.data.frame(out)
+  row.names(out) <- NULL
+
+  attributes(out) <- c(attributes(out), attri)
+  class(out) <- c("performance_model", class(out))
+
+  out
+}
+
+.get_sigma_bfbayesfactor <- function(model, average = FALSE, prior_odds = NULL) {
+  if (average) {
+    return(.get_sigma_bfbayesfactor_model_average(model, prior_odds = prior_odds))
+  }
+
+  params <- insight::get_parameters(model)
+  if (!"sig2" %in% colnames(params)) stop("This is not a linear model.")
+  sqrt(params$sig2)
+}
+
+
+.get_sigma_bfbayesfactor_model_average <- function(model, prior_odds = NULL) {
+  if (!requireNamespace("BayesFactor", quietly = TRUE)) {
+    stop("Package `BayesFactor` needed for this function to work. Please install it.")
+  }
+
+  BFMods <- bayestestR::bayesfactor_models(model, verbose = FALSE)
+
+  # extract parameters
+  intercept_only <- which(BFMods$Model == "1")
+  params <- vector(mode = "list", length = nrow(BFMods))
+  for (m in seq_along(params)) {
+    if (length(intercept_only) && m == intercept_only) {
+      y <- insight::get_response(model)
+      params[[m]] <- rep(sd(y), 4000)
+    } else if (m == 1) {
+      # If the model is the "den" model
+      params[[m]] <- suppressMessages(.get_sigma_bfbayesfactor(1 / model[1]))
+    } else {
+      params[[m]] <- suppressMessages(.get_sigma_bfbayesfactor(model[m - 1]))
+    }
+  }
+
+  params <- lapply(params, data.frame)
+
+
+  # Compute posterior model probabilities
+  if (!is.null(prior_odds)) {
+    prior_odds <- c(1, prior_odds)
+  } else {
+    prior_odds <- rep(1, nrow(BFMods))
+  }
+  posterior_odds <- prior_odds * BFMods$BF
+  posterior_odds <- posterior_odds[-1] / posterior_odds[1]
+
+  do.call(bayestestR::weighted_posteriors,
+          c(params, list(missing = 0, prior_odds = posterior_odds)))[[1]]
+}